introduzione pk
introdotti pk e aggiunta nel main richiesta numero cluster e numero di esecuzione
