+++ src/MainTest.java
@@ -1,3 +1,10 @@
+import data.Data;
+import mining.KMeansMiner;
+
+import java.util.Scanner;
+
+import static keyboardinput.Keyboard.*;
+
 
 public class MainTest {
 
@@ -5,17 +12,33 @@ public class MainTest {
 	 * @param args
 	 */
 	public static void main(String[] args) {
-		
-		
+
 		Data data = new Data();
 		System.out.println(data);
-		int k = 3;
-		KMeansMiner kmeans = new KMeansMiner(k);
-		int numIter = kmeans.kmeans(data);
-		System.out.println("Numero di Iterazione:"+numIter);
-		System.out.println(kmeans.getC().toString(data));
-		
-		
-	}
 
+		char choice;
+		do {
+			int k = 0;
+			do {
+				System.out.print("Enter the number : ");
+				k = readInt();
+			} while (k <= 0);
+
+			KMeansMiner kmeans = new KMeansMiner(k);
+			int numIter = kmeans.kmeans(data);
+			System.out.println("Numerous di Iterazione:" + numIter);
+			System.out.println(kmeans.getC().toString(data));
+
+
+			do {
+				System.out.print("Do you want to continue running? : ");
+				System.out.print("For YES type Y, for NO type N : ");
+				choice = readChar();
+			} while ((choice == 'Y') || (choice == 'N'));
+		}while(choice =='Y');
+
+
+
+
+	}
 }
