+++ src/data/Tuple.java
@@ -1,3 +1,5 @@
+package data;
+
 public class Tuple {
     private Item[] tuple;
 
@@ -5,11 +7,11 @@ public class Tuple {
         tuple = new Item[size];
     }
 
-    int getLength(){
+    public int getLength(){
         return tuple.length;
     }
 
-    Item get(int i){
+    public Item get(int i){
         return tuple[i];
     }
 
@@ -17,7 +19,7 @@ public class Tuple {
         tuple[i] = c;
     }
 
-    double getDistance(Tuple obj) {
+    public double getDistance(Tuple obj) {
         int length;
         double distance = 0;
 
@@ -28,7 +30,7 @@ public class Tuple {
         return distance;
     }
 
-    double avgDistance(Data data, int[] clusteredData) {
+    public double avgDistance(Data data, int[] clusteredData) {
         double p = 0.0, sumD = 0.0;
         for(int i = 0; i < clusteredData.length; i++) {
             double d = getDistance(data.getItemSet(clusteredData[i]));
