+++ src/mining/Cluster.java
@@ -1,10 +1,15 @@
+package mining;
+
+import data.Data;
+import data.Tuple;
+import utility.ArraySet;
 
 public class Cluster {
 	private Tuple centroid;
 
-	private ArraySet clusteredData; 
+	private ArraySet clusteredData;
 	
-	/*Cluster(){
+	/*mining.Cluster(){
 		
 	}*/
