+++ src/mining/ClusterSet.java
@@ -1,4 +1,8 @@
-import java.util.Arrays;
+package mining;
+
+import data.Data;
+import data.Tuple;
+
 
 public class ClusterSet {
     private Cluster[] C;
