+++ src/utility/ArraySet.java
@@ -1,17 +1,19 @@
+package utility;
+
 import java.util.Arrays;
 
 
 public class ArraySet {
 	private boolean set[];
 	private int size=0;
-	ArraySet (){
+	public ArraySet(){
 		set=new boolean[50];
 		for(int i = 0; i<set.length; i++)
 			set[i]=false;
 	}
 	
 	//return true if add is changing the arraySet
-	boolean add(int i){
+	public boolean add(int i){
 		if(i>=set.length)
 		{
 			//enlarge the set
@@ -30,7 +32,7 @@ public class ArraySet {
 	}
 
 	// cancella un true se c'è in posizione i e ritorna il risultato dell'operazione
-	boolean delete(int i){
+	public boolean delete(int i){
 		if(i<size){
 			boolean deleted=set[i];
 			set[i]=false;
@@ -48,12 +50,12 @@ public class ArraySet {
 	}
 
 	// ritorna il valore contentuto nella posizione i
-	boolean get(int i){
+	public boolean get(int i){
 		return set[i];
 	}
 
 	// ritorna un array di interi con le posizioni in cui si trova un "true"
-	int[] toArray(){
+	public int[] toArray(){
 		int a[]=new int[0];
 		for(int i=0;i<size;i++){
 			if(get(i)){
